serving/torchserve/requirements.txt
ultralytics
pillow
numpy
